% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mat.R
\name{blocks_names}
\alias{blocks_names}
\title{blocks_names}
\usage{
blocks_names(named_colorvector, border_color = NA, label_size = 0.4)
}
\arguments{
\item{named_colorvector}{named vector containing colors}

\item{border_color}{border color of boxes}

\item{label_size}{relative size of labels}
}
\description{
creates boxes around text labels with colors indicated in named_colorvector
}
