% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{clustering-list}
\alias{clustering-list}
\title{clustering}
\description{
clustering
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{$Embryos}
\keyword{$Genes}
\keyword{Alphabetical}
\keyword{ClusterNo.}
\keyword{Clusters}
\keyword{E_ID.}
\keyword{Gene}
\keyword{GeneName.}
\keyword{IDs}
\keyword{Names}
\keyword{Numerical}
\keyword{\\item}
\keyword{\item}
\keyword{\itemize{}
\keyword{be}
\keyword{can}
\keyword{clustering}
\keyword{code}
\keyword{data}
\keyword{embryos}
\keyword{figure2}
\keyword{for}
\keyword{function}
\keyword{genes}
\keyword{heatmap}
\keyword{in}
\keyword{information.}
\keyword{list}
\keyword{marker}
\keyword{object}
\keyword{of}
\keyword{recomputed}
\keyword{sequenced}
\keyword{the}
\keyword{via}
\keyword{with}
\keyword{}}
