% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_clustering.R
\name{clustering_kmeans}
\alias{clustering_kmeans}
\title{clustering_kmeans}
\usage{
clustering_kmeans(
  exprmat = exprmat_TMM_minmax,
  selection = markergenes$GeneName,
  embryo_i = embryo_info,
  embryo_ID_column = "E_ID",
  gene_ID_column = "GeneName",
  seed = 123,
  rowkm = 6,
  columnkm = 4,
  rowkm_rep = 800,
  columnkm_rep = 800,
  save = F,
  output = file.path("data", "clusterlist.rda")
)
}
\arguments{
\item{exprmat}{matrix with data to base clustering on (gene names should be row names)}

\item{selection}{selection of genes to cluster}

\item{embryo_i}{matrix with info on embryos}

\item{embryo_ID_column}{column name for identifiers for embryos}

\item{gene_ID_column}{column name for identifiers for genes}

\item{seed}{seed for kmeans}

\item{rowkm}{number of clusters in rows}

\item{columnkm}{number of clusters in columns}

\item{save}{save a list object with clustering}

\item{output}{output file for save}

\item{row_km_repeats}{number of repeats for rows for kmeans}

\item{column_km_repeats}{number of repeats for columns for kmeans}
}
\description{
get kmeans clusters for genes
}
