% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{cmtopixel}
\alias{cmtopixel}
\title{cmtopixel}
\usage{
cmtopixel(cm, res = 600)
}
\arguments{
\item{cm}{cm to convert}

\item{res}{resolution}
}
\description{
converts cm to pixelvalues at res dpi
}
