% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{colorscheme-data}
\alias{colorscheme-data}
\title{colorscheme-df}
\description{
colorscheme-df
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{assignments}
\keyword{color}
\keyword{containing}
\keyword{data}
\keyword{dataframe}
