% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plots.R
\name{corr_groups}
\alias{corr_groups}
\title{corr_groups}
\usage{
corr_groups(
  expr_df,
  corr_type = "spearman",
  R_limit = 0.7,
  p_limit = 0.05,
  h_cut = 0.7,
  n_group = NULL
)
}
\arguments{
\item{expr_df}{dataframe with gene expression (columnnames must be genes)}

\item{corr_type}{type of correlation to use in rcorr (spearman, pearson, ...)}

\item{R_limit}{R/Rho threshold}

\item{p_limit}{p value threshold}

\item{h_cut}{height to cut the hierarchical clustering tree}

\item{n_group}{number of groups (NULL if group number should not be pre-determined)}
}
\description{
form groups with high correlations
}
