% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plots.R
\name{correval_plts}
\alias{correval_plts}
\title{correval_plts}
\usage{
correval_plts(
  expr_mat = exprmat_TMM_minmax,
  gene_annot = markergenes,
  gene_ID_column = "GeneName",
  corr_type = "spearman",
  coef_name = "rho",
  R_limit = 0.7,
  p_limit = 0.05,
  h_cut = 0.7,
  n_group = NULL,
  labels = "AUTO"
)
}
\arguments{
\item{expr_mat}{matrix with gene expression (genes as rows)}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).}

\item{corr_type}{type of correlation to use in rcorr (spearman, pearson, ...)}

\item{coef_name}{name of coefficient in plots}

\item{R_limit}{R/Rho threshold}

\item{p_limit}{p value threshold}

\item{h_cut}{height to cut the hierarchical clustering tree}

\item{n_group}{number of groups (NULL if group number should not be pre-determined)}

\item{labels}{vector with labels (AUTO/auto for big/small automatic letters)}
}
\description{
plots gene groups with high correlation
}
