% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mat.R
\name{corrmat}
\alias{corrmat}
\title{corrmat}
\usage{
corrmat(
  expr_mat = exprmat_TMM_minmax,
  gene_annot = markergenes,
  gene_ID_column = "GeneName",
  colors = colorscheme,
  colors_chosen = "Compartment",
  label_size = 0.4
)
}
\arguments{
\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).}

\item{colors}{dataframe with color assignement}

\item{colors_chosen}{refers to Item column in colors to be display (Process or Compartment)}

\item{label_size}{size of labels relative to default}

\item{expr_df}{vector of selected genes}
}
\description{
Plots expression levels of selected df
}
