% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embryo_size_time.R
\name{embryo_cell_time_alt}
\alias{embryo_cell_time_alt}
\title{embryo_cell_time_alt}
\usage{
embryo_cell_time_alt(
  e_data = embryo_info,
  xparam = "Days_since_1.8.",
  coloring = "ClusterNo",
  shapes = "Morphology",
  cells_on_x = F,
  colors = colorscheme,
  colorclass = "EmbryoCluster",
  xlabel = bquote("Days from " ~ 1^"st" ~ "of August"),
  legend_pos = c(0.05, 0.95)
)
}
\arguments{
\item{e_data}{data frame to use}

\item{xparam}{column to plot on x axis}

\item{coloring}{column to use for coloring}

\item{cells_on_x}{display cells on x axis (T/F) if == T plot is flipped}

\item{colors}{data frame with colorscheme}

\item{colorclass}{class in colors to use}

\item{xlabel}{label of x axis}

\item{legend_pos}{position of legend (vector of lenght 2, with values from 0 to 1)}
}
\description{
plots the embryo cells against time (alternative with including cluster)
}
