% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{embryo_hm_annot}
\alias{embryo_hm_annot}
\title{embryo_hm_annot}
\usage{
embryo_hm_annot(
  expr_mat,
  clusterEmbryos = clustering$Embryos,
  ClusterIDcolumnE = "ClusterNo",
  embryo_annot = embryo_info,
  embryo_ID_column = "E_ID",
  cell_col_name = "Log2Cells",
  morph_col_name = "Morphology",
  colors_cells = (colorscheme \%>\% filter(Class == "Cells"))$Color,
  colors_morph = colorscheme \%>\% filter(Class == "Morphology"),
  colors_cluster = (colorscheme \%>\% filter(Class == "EmbryoCluster")),
  RC = "col",
  show_lg = TRUE,
  anno_size = unit(2, "mm"),
  show_name = FALSE,
  base_size = 8,
  return_cellno = F,
  return_morph = F,
  return_cluster = F
)
}
\arguments{
\item{expr_mat}{numeric matrix containing expression data}

\item{clusterEmbryos}{data frame containing column with embryo identifier (embryo_ID_column) and cluster assignment}

\item{ClusterIDcolumnE}{column name in clusterEmbryos with names for embryo / sample clusters}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_ID_column}{column name with embryo / sample ID variable (must be identical to either column or row names of expr_mat)}

\item{cell_col_name}{column name in embryo_annot containing log2-transformed cell numbers}

\item{morph_col_name}{column name in embryo_annot containing log2-transformed cell numbers}

\item{colors_cells}{color vector with 3 elements for cell numbers}

\item{colors_morph}{data frame containing color assignments for morphologies}

\item{colors_cluster}{data frame containing color assignments for cluster}

\item{RC}{direction of annotation ("col" or "row")}

\item{show_lg}{display annotation legend (T/F)}

\item{anno_size}{size of annotation (unit())}

\item{base_size}{base size for fonts}

\item{return_cellno}{return annotation for cell number? (T/F)}

\item{return_morph}{return annotation for morphology? (T/F)}

\item{return_cluster}{return annotation for clustering? (T/F)}

\item{Embryoskm}{number of embryo / sample clusters for kmeans clustering}

\item{Embryoskm_repeats}{number of repeats for kmeans clustering for embryo / sample clusters}
}
\description{
heatmap annotations for embryos
}
