% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{embryo_info-data}
\alias{embryo_info-data}
\title{embryo_info-df}
\description{
embryo_info-df
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{containing}
\keyword{data}
\keyword{dataframe}
\keyword{embryos}
\keyword{for}
\keyword{in}
\keyword{included}
\keyword{metadata}
\keyword{study}
