% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiblast_mophology.R
\name{epiblast_morphology}
\alias{epiblast_morphology}
\title{epiblast_morphology}
\usage{
epiblast_morphology(
  ki_data = ki67evaluation,
  filter_col = "Stage",
  filter_var = "Diapausing",
  grouping_var = "ICM_shape",
  embryo_annot = embryo_info,
  x_var = "Nuclei_Tot",
  y_var = "Days_since_1.8.",
  colors = colorscheme[colorscheme$Class == "ICM", ],
  base_size = 8
)
}
\arguments{
\item{ki_data}{data frame to use}

\item{filter_col}{column in ki_data to use for filtering data}

\item{filter_var}{variable(s) to filter ki_data for}

\item{grouping_var}{column in ki_data containing the groups}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{x_var}{variable in ki_data to plot on x axis}

\item{y_var}{variable in ki_data to plot on y axis}

\item{colors}{dataframe with color assignement}

\item{base_size}{base size for text (pt)}
}
\value{
ggplot object
}
\description{
plots epiblast morphology and stat-boxes
}
