% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_patterns.R
\name{expr_pattern_selected}
\alias{expr_pattern_selected}
\title{expr_pattern_selected}
\usage{
expr_pattern_selected(
  expr_mat = exprmat_TMM_minmax,
  gene_as_rows = T,
  embryo_annot = embryo_info,
  embryo_ID_column = "E_ID",
  embryo_plot_var = "Cells",
  gene_annot = markergenes,
  gene_ID_column = "GeneName",
  grouping_var = "show_pattern",
  sorting_var = "plot_no",
  color_var = "Compartment",
  base_size = 8,
  colors = colorscheme,
  ylab_txt = "scaled TMM",
  xlab_txt = "Cells [#]"
)
}
\arguments{
\item{expr_mat}{numeric matrix containing expression data, colnames (or rownames) should contain values found in embryo_ID_column}

\item{gene_as_rows}{are genes as rows? (T/F) --> if false matrix will by transposed}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_ID_column}{column name with embryo / sample ID variable (must be identical to either column or row names of expr_mat)}

\item{embryo_plot_var}{variable in embryo_annot to plot (numeric)}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).
Order of elements in gene_ID_column will be maintained inside each gene cluster.}

\item{grouping_var}{variable in gene_annot containing the groups to be displayed}

\item{sorting_var}{variable in gene_annot containing the order of plots to be displayed}

\item{color_var}{variable in gene_annot to use for coloring (must be "Class" in colors)}

\item{base_size}{base size for text (pt)}

\item{colors}{dataframe with color assignement}

\item{ylab_txt}{text on y axis}

\item{xlab_txt}{text on x axis}
}
\description{
expr_pattern_selected
}
