% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_patterns.R
\name{expr_vs_cells}
\alias{expr_vs_cells}
\title{expr_vs_cells}
\usage{
expr_vs_cells(
  gene = "NANOG",
  expr_mat = exprmat_TMM_minmax,
  embryo_annot = embryo_info,
  embryo_plot_var = "Cells",
  gene_annot = markergenes,
  base_size = 8,
  vertical_lines = TRUE,
  points = FALSE,
  ylab_txt = "scaled TMM",
  xlab_txt = "Cells [#]",
  linecol = "black",
  filling = "black"
)
}
\arguments{
\item{expr_mat}{numeric matrix containing expression data, colnames (or rownames) should contain values found in embryo_ID_column}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_plot_var}{column in embryo annot to plot on x}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{base_size}{base size for text (pt)}

\item{vertical_lines}{display vertical lines as orientation}

\item{points}{display individual data points}

\item{ylab_txt}{text on y axis}

\item{xlab_txt}{text on x axis}

\item{linecol}{color for lines}

\item{filling}{= color for shades (CI)}

\item{gene_as_rows}{are genes as rows? (T/F) --> if false matrix will by transposed}
}
\description{
plots gene expression vs embryo cell number
}
