% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressiondata_rda.R
\name{expression_data_rda}
\alias{expression_data_rda}
\title{expression_data_rda}
\usage{
expression_data_rda(
  data_location = file.path("extdata", "raw_tables"),
  prefix = "RoeDeerEmbryo_counts_",
  inputtypes = c("expected", "TPM", "FPKM", "TMM"),
  suffix = ".txt",
  save_list = F,
  scaled_counts = F,
  scaling_methods = c("minmax", "zscore"),
  outdir = "data",
  outfilename = "exprmat_",
  return_list = F,
  rem_outlier = T
)
}
\arguments{
\item{data_location}{folder path containing expression data, default is file.path("data", "raw_tables")}

\item{prefix}{file name prior to input type, default is "RoeDeerEmbryo_counts_"}

\item{inputtypes}{vector with types of input (must be in filename before suffix), default is c("expected", "TPM", "FPKM")}

\item{suffix}{suffix of input file, default is ".txt"}

\item{save_list}{save the the a list as Rds file? (T/F)}

\item{scaled_counts}{also create scaled list for each normalization, default = F}

\item{scaling_methods}{which scaling function to use, default = c("minmax", "zscore")}

\item{outdir}{directory for output}

\item{outfilename}{file name for output (without suffix)}

\item{return_list}{return the generated list}

\item{rem_outlier}{remove one embryo with low quality reads? (T/F)}
}
\value{
list object with differently normalized and/or scaled expression data
}
\description{
create an Rds object or a list with all types of data matrices.
}
