% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_S1_expression_levels.R
\name{expression_swarm_plt}
\alias{expression_swarm_plt}
\title{expression_swarm_plt}
\usage{
expression_swarm_plt(
  expr_mat = exprmat_TMM,
  gene_annot = markergenes,
  gene_ID_column = "GeneName",
  colors = colorscheme,
  colors_chosen = "Compartment",
  sorting_col = "Compartment_No",
  logscale = T,
  allin1 = F
)
}
\arguments{
\item{expr_mat}{matrix with expression data (not scaled!)}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).}

\item{colors}{dataframe with color assignement}

\item{colors_chosen}{refers to Item column in colors to be display (Process or Compartment)}

\item{sorting_col}{column to sort plot by}

\item{logscale}{do you want a log-y axis ? (T/F)}

\item{allin1}{should all plots be in one ? (T/F) (if F will generate four plots via quantiles)}
}
\description{
Plots expression levels of selected df
}
