% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{exprmat_TMM-data}
\alias{exprmat_TMM-data}
\title{exprmat_TMM-mat}
\description{
exprmat_TMM-mat
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{(unscaled)}
\keyword{TMM}
\keyword{data}
\keyword{expression}
\keyword{matrix}
\keyword{normalized}
\keyword{with}
