% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.R
\name{figure1}
\alias{figure1}
\title{figure1}
\usage{
figure1(image_dir = file.path("extdata", "images"), fontsize_labels = 12)
}
\arguments{
\item{image_dir}{directory with images}

\item{fontsize_labels}{size of labels (pt)}

\item{h}{plot height}

\item{w}{plot width}

\item{save}{save plot? (T/F)}
}
\value{
ggplot object
}
\description{
generates figure 1 for the manuscript
}
