% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{figure2}
\usage{
figure2(
  expr_mat = exprmat_TMM_minmax,
  gene_as_rows = T,
  compute_clusters_new = F,
  seed = 123,
  embryo_annot = embryo_info,
  embryo_ID_column = "E_ID",
  clusterEmbryos = clustering$Embryos,
  ClusterIDcolumnE = "ClusterNo",
  Embryoskm = 4,
  Embryoskm_repeats = 800,
  gene_annot = markergenes,
  gene_ID_column = "GeneName",
  clusterGenes = clustering$Genes,
  ClusterIDcolumnG = "ClusterNo",
  Genekm = 6,
  Genekm_repeats = 800,
  grouping_var = "Compartment",
  base_size = 8,
  colors = colorscheme,
  scaling = "minmax",
  show_expr_lvls = T,
  value_name = "scaled\\nTMM\\ncounts",
  filter_data = F,
  unscaled_mat = exprmat_TMM,
  min_TMM = 1,
  min_s = 3
)
}
\arguments{
\item{expr_mat}{numeric matrix containing expression data}

\item{gene_as_rows}{are genes as rows? (T/F) --> if false matrix will by transposed}

\item{compute_clusters_new}{re-compute clusters for embryos / samples and genes}

\item{seed}{seed used for kmeans clustering}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_ID_column}{column name with embryo / sample ID variable (must be identical to either column or row names of expr_mat)}

\item{clusterEmbryos}{data frame with cluster assignment for embryos / samples (ignored if compute_clusters_new = T)}

\item{ClusterIDcolumnE}{column name in clusterEmbryos with names for embryo / sample clusters}

\item{Embryoskm}{number of embryo / sample clusters for kmeans clustering}

\item{Embryoskm_repeats}{number of repeats for kmeans clustering for embryo / sample clusters}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).
Order of elements in gene_ID_column will be maintained inside each gene cluster.}

\item{clusterGenes}{data frame with cluster assignment for genes (ignored if compute_clusters_new = T)}

\item{ClusterIDcolumnG}{column name in clusterGenes with names for gene clusters}

\item{Genekm}{number of genee clusters for kmeans clustering}

\item{Genekm_repeats}{number of repeats for kmeans clustering for gene clusters}

\item{grouping_var}{variable in gene_annot containing the groups to be displayed (names must be found in coloring)}

\item{base_size}{base size for text (pt)}

\item{colors}{dataframe with color assignement}

\item{scaling}{was data scaled minmax or zscore (will change colors of heatmap)}

\item{show_expr_lvls}{should absolute expression levels be shown (T/F)? If T, provide unscaled_mat}

\item{value_name}{name for heatmap variable}

\item{filter_data}{should data be filtered (T/F)? If T, provide unscaled_mat, min_TMM, and min_s.}

\item{unscaled_mat}{dataframe with unscaled expression values (identical format as expr_mat)}

\item{min_TMM}{min TMM value that needs to be observed in at least min_s samples for a gene to be shown}

\item{min_s}{min number of samples that need to display at least TMM value for a gene to be shown}
}
\description{
plots heatmap using ComplexHeatmap
}
