% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{genes_hm_annot}
\alias{genes_hm_annot}
\title{genes_hm_annot}
\usage{
genes_hm_annot(
  expr_mat,
  clusterGenes = clustering$Genes,
  ClusterIDcolumnG = "ClusterNo",
  gene_ID_column = "GeneName",
  gene_annot = markergenes,
  grouping_var = "Compartment",
  sorting_var = "Compartment_No",
  colors = colorscheme,
  RC = "row",
  show_lg = TRUE,
  anno_size = unit(4, "mm"),
  show_name = FALSE,
  base_size = 8
)
}
\arguments{
\item{expr_mat}{numeric matrix containing expression data}

\item{clusterGenes}{data frame containing column with Gene identifier (GeneName) and cluster assignment (ABC)}

\item{ClusterIDcolumnG}{column name in clusterGenes with names for gene clusters}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{grouping_var}{variable in gene_annot containing the groups to be displayed (names must be found in coloring)}

\item{sorting_var}{variable in gene_annot for which legend should be sorted}

\item{colors}{data frame with colorscheme}

\item{RC}{direction of annotation ("col" or "row")}

\item{show_lg}{display annotation legend (T/F)}

\item{anno_size}{size of annotation (unit())}

\item{show_name}{show annotation name (T/F)}

\item{base_size}{base size for fonts}
}
\description{
heatmap annotations for embryos
}
