% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki67_barplot.R
\name{ki67_barplot}
\alias{ki67_barplot}
\title{ki67_barplot}
\usage{
ki67_barplot(
  ki_data = ki67evaluation,
  grouping = "Stage",
  value_column = "perc_tot_ki",
  ylabel = "\% of Ki67 positive Nuclei",
  colors = colorscheme,
  colorclass = "Morphology"
)
}
\arguments{
\item{ki_data}{data frame to use}

\item{grouping}{column of ki_data containing grouping variable}

\item{value_column}{column of ki_data containing value to display}

\item{ylabel}{label of y axis}

\item{colors}{data frame with colorscheme}

\item{colorclass}{class in colors to use}
}
\description{
plots barplot of ki67 counting data
}
