% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ki67evaluation-data}
\alias{ki67evaluation-data}
\title{ki67evaluation-df}
\description{
ki67evaluation-df
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{Ki67}
\keyword{containing}
\keyword{counting}
\keyword{data}
\keyword{dataframe}
\keyword{for}
\keyword{stainings}
