% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mat.R
\name{light_dark}
\alias{light_dark}
\title{light_dark}
\usage{
light_dark(hexcolor)
}
\arguments{
\item{hexcolor}{color in hex}
}
\description{
returns black for light colors and white for dark colrs to allow better readability
}
