% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{markergenes-data}
\alias{markergenes-data}
\title{markergenes-df}
\description{
markergenes-df
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{assignment}
\keyword{containing}
\keyword{data}
\keyword{dataframe}
\keyword{genes}
\keyword{groups}
\keyword{into}
\keyword{marker}
\keyword{of}
