% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_rda.R
\name{metadata_rda}
\alias{metadata_rda}
\title{metadata_rda}
\usage{
metadata_rda(
  data_location = file.path("extdata", "raw_tables", "metadata"),
  suffix = ".xlsx",
  returnlist = F,
  outdir = "data"
)
}
\arguments{
\item{data_location}{folder path containing metadata files, default is file.path("data", "raw_tables", "metadata")}

\item{suffix}{suffix of input file, default is ".xlsx"}

\item{returnlist}{return the list? (T/F)}

\item{outdir}{directory for output}
}
\value{
list object with differently normalized and/or scaled expression data
}
\description{
creates an rda objects of all metadata
}
