% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mat.R
\name{mycolor_label}
\alias{mycolor_label}
\title{mycolor_label}
\usage{
mycolor_label(
  color_annotation,
  label_size = 0.8,
  point_size = 0.5,
  x_shift = -6,
  y_shift = 2,
  y_distance = 1.5
)
}
\arguments{
\item{color_annotation}{color annotation df}

\item{label_size}{relative size of labels}

\item{point_size}{length of the sides of the squares}

\item{x_shift}{shift of legend in x direction}

\item{y_shift}{shift of legend in y direction}

\item{y_distance}{distance between legend lines}
}
\description{
creates custom color label legend
}
