% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_helper.R
\name{p_value_df}
\alias{p_value_df}
\title{p_value_df}
\usage{
p_value_df(df, variable, grouping, log_scale = TRUE)
}
\arguments{
\item{df}{data frame to create p_value data frame from}

\item{variable}{variable to analyze}

\item{grouping}{grouping variable}

\item{log_scale}{should data be log_tranformed (T/F)}
}
\description{
return df with p value
}
