% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(
  expr_mat = exprmat_TMM_minmax,
  gene_as_rows = T,
  embryo_annot = embryo_info,
  embryo_ID_column = "E_ID",
  clusterEmbryos = clustering$Embryos,
  ClusterIDcolumnE = "ClusterNo",
  gene_annot = markergenes,
  gene_ID_column = "GeneName",
  grouping_var = "Compartment",
  colorsE = colorscheme[colorscheme$Class == "EmbryoCluster", ][["Color"]],
  colorsG = colorscheme[colorscheme$Class == "Compartment", ][["Color"]],
  base_size = 8,
  colorlab = "Cluster\\nNumber",
  output = "ind"
)
}
\arguments{
\item{expr_mat}{numeric matrix containing expression data, colnames (or rownames) should contain values found in embryo_ID_column}

\item{gene_as_rows}{are genes as rows? (T/F) --> if false matrix will by transposed}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_ID_column}{column name with embryo / sample ID variable (must be identical to either column or row names of expr_mat)}

\item{clusterEmbryos}{data frame with cluster assignment for embryos / samples (ignored if compute_clusters_new = T)}

\item{ClusterIDcolumnE}{column name in clusterEmbryos with names for embryo / sample clusters}

\item{gene_annot}{data frame containing annotations for genes (also defines which genes to display)}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).
Order of elements in gene_ID_column will be maintained inside each gene cluster.}

\item{grouping_var}{variable in gene_annot containing the groups to be displayed}

\item{colorsE}{vector with colors for embryo clusters}

\item{colorsG}{vector with colors for genes}

\item{base_size}{base size for text (pt)}

\item{output}{which plot(s) to return c("ind", biplot","both")}
}
\description{
plot_pca
}
