% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppw_hub_heatmap.R
\name{pppw_heatmap}
\alias{pppw_heatmap}
\title{metadata_rda}
\usage{
pppw_heatmap(
  pp_pws = pppw_hubgenes,
  gene_ID_column = "GeneName",
  grouping_var = "Pathway",
  expr_mat = exprmat_TMM_minmax,
  embryo_annot = embryo_info,
  embryo_ID_column = "E_ID",
  clusterEmbryos = clustering$Embryos,
  ClusterIDcolumnE = "ClusterNo",
  orderE = c("cluster", "cells"),
  colors = colorscheme,
  filter_data = F,
  unscaled_mat = exprmat_TMM,
  min_TMM = 1,
  min_s = 3,
  base_size = 8
)
}
\arguments{
\item{pp_pws}{list with gene names and pathway association}

\item{gene_ID_column}{column name with gene ID (must be identical to either column or row names of expr_mat).}

\item{grouping_var}{variable in pp_pws containing the groups to be displayed}

\item{expr_mat}{numeric matrix containing expression data}

\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_ID_column}{column name with embryo / sample ID variable (must be identical to either column or row names of expr_mat)}

\item{clusterEmbryos}{data frame with cluster assignment for embryos / samples (ignored if compute_clusters_new = T)}

\item{ClusterIDcolumnE}{column name in clusterEmbryos with names for embryo / sample clusters}

\item{orderE}{order of embryos (cluster (precomputed based on figure 2 heatmap) or cells)}

\item{colors}{dataframe with color assignement}

\item{filter_data}{should data be filtered (T/F)? If T, provide unscaled_mat, min_TMM, and min_s.}

\item{unscaled_mat}{dataframe with unscaled expression values (identical format as expr_mat)}

\item{min_TMM}{min TMM value that needs to be observed in at least min_s samples for a gene to be shown}

\item{min_s}{min number of samples that need to display at least TMM value for a gene to be shown}

\item{base_size}{base size for text (pt)}
}
\value{
heatmap object
}
\description{
creates an rda objects of all metadata
}
