% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pppw_hubgenes-data}
\alias{pppw_hubgenes-data}
\title{pppw_hubgenes-df}
\description{
pppw_hubgenes-df
}
\references{
\url{to be uploaded}
}
\author{
Anna Rueegg \email{anna.rueegg@usys.ethz.ch}
}
\keyword{assignment}
\keyword{containing}
\keyword{data}
\keyword{dataframe}
\keyword{for}
\keyword{hubgenes}
\keyword{of}
\keyword{pathways}
\keyword{pluripotency}
