% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_funs.R
\name{save_heatmap}
\alias{save_heatmap}
\title{save_heatmap}
\usage{
save_heatmap(
  w = 18,
  h = 22,
  outdir = "output",
  filename = "Heatmap",
  outformat = "svg",
  ht = custom_heatmap(),
  resolution = 600,
  seed = 123
)
}
\arguments{
\item{w}{width in cm}

\item{h}{height in cm}

\item{outdir}{output directory}

\item{filename}{add custom filename}

\item{outformat}{format for output (svg, pdf, png)}

\item{ht}{heatmap to print or save}

\item{resolution}{resolution (dpi)}

\item{seed}{seed used for kmeans clustering}
}
\description{
draws heatmap with the option to save it.
}
