% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_funs.R
\name{save_plt}
\alias{save_plt}
\title{save_plt}
\usage{
save_plt(
  plt = ggplot(),
  w = 18,
  h = 22,
  outdir = "output",
  filename = "",
  outformat = "pdf",
  resolution = 600
)
}
\arguments{
\item{plt}{plot to save}

\item{w}{width in cm}

\item{h}{height in cm}

\item{outdir}{output directory}

\item{filename}{add custom filename}

\item{outformat}{format for output (svg, pdf, png)}

\item{resolution}{resolution (dpi)}
}
\description{
plots and saves any plot
}
