% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_funs.R
\name{scale_df}
\alias{scale_df}
\title{scale_df}
\usage{
scale_df(df, fun = minmax, direction = "cols")
}
\arguments{
\item{df}{a numeric dataframe}

\item{fun}{function to be applied}

\item{direction}{direction in which the function will be applied. Either rows or cols, default = cols.}
}
\description{
performs a min-max standardization of a dataframe, accross one direction
}
