% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_E_cluster.R
\name{stats_E_cluster}
\alias{stats_E_cluster}
\title{stats_E_cluster}
\usage{
stats_E_cluster(
  embryo_annot = embryo_info,
  embryo_ID_column = "E_ID",
  selection_var = "Cells",
  color_var = "Morphology",
  fill_var = "ClusterNo",
  clusterEmbryos = clustering$Embryos,
  ClusterIDcolumnE = "ClusterNo",
  log_axis = T,
  colors = colorscheme,
  base_size = 8,
  horizontal_legend = TRUE,
  xlab_txt = "Embryo Cluster No.",
  ylab_txt = "Cells [#]",
  color_lab_txt = "Morphology",
  fill_lab_txt = "Embryo\\nCluster No.",
  grouponY = T
)
}
\arguments{
\item{embryo_annot}{data frame containing annotations for embryo data}

\item{embryo_ID_column}{column name with embryo / sample ID variable (must be identical to either column or row names of expr_mat)}

\item{selection_var}{variable in embryo_annot to analyze (numeric only)}

\item{color_var}{variable in embryo_annot to use for coloring points}

\item{fill_var}{variable in embryo_annot or clusterEmbryos to use for filling violoin plots}

\item{clusterEmbryos}{data frame with cluster assignment for embryos / samples (ignored if compute_clusters_new = T)}

\item{ClusterIDcolumnE}{column name in clusterEmbryos with names for embryo / sample clusters}

\item{log_axis}{plot on logarithmized axis}

\item{colors}{dataframe with color assignement}

\item{base_size}{base size for text (pt)}

\item{horizontal_legend}{plot legend horizontally}

\item{xlab_txt}{text on x axis}

\item{ylab_txt}{text on y axis}

\item{color_lab_txt}{text for color legend}

\item{fill_lab_txt}{text for fill legend}

\item{grouponY}{plot group on y axis}
}
\description{
violin plots of embryo clusters (cells, days)
}
